requirements.txt
spotipy==2.23.0
Pillow==10.0.0
requests==2.31.0
PyQt5==5.15.9
loguru==0.7.2